COM-1755 using mixin to retrieve learners
